Move help and error modal components to modals folder
