+++ src/components/modals/ErrorModal.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { computed } from 'vue';
-import { useUiStore } from '../stores/uiStore';
+import { useUiStore } from '../../stores/uiStore';
 
 const uiStore = useUiStore();
